+++ client/internal/view/item/text/text.go
@@ -48,9 +48,18 @@ func (v *View) Init() (common.KeyHandlerFnc, common.Help) {
 		AddButton("Save", func() {
 			go v.save()
 		})
+	modal := tview.NewModal().
+		SetText("Are you sure?").
+		AddButtons([]string{"Yes", "No"}).
+		SetDoneFunc(func(buttonIndex int, buttonLabel string) {
+			if buttonIndex == 0 {
+				go v.delete()
+			}
+			v.Frame.SetPrimitive(form)
+		})
 	if v.item.ID != "" {
 		form.AddButton("Delete", func() {
-			go v.delete()
+			v.Frame.SetPrimitive(modal)
 		})
 	}
 	form.SetBorderPadding(1, 1, 0, 1)
